+++ examples/demo_handover_env.py
@@ -18,7 +18,7 @@ def main():
     while True:
         env.reset(scene_id=scene_id)
         for _ in range(int(3.0 / cfg.SIM.TIME_STEP)):
-            action = np.array(cfg.ENV.PANDA_INITIAL_POSITION, dtype=np.float32)
+            action = np.array(cfg.ENV.PANDA_INITIAL_POSITION)
             action += np.random.uniform(low=-1.0, high=+1.0, size=len(action))
             env.step(action)
